backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
pandas>=2.1.0
scikit-learn>=1.3.0
pytest-django>=4.7.0
pytest>=7.4.0
prophet>=1.1.5
statsmodels>=0.14.0
celery[redis]>=5.3.0
django-redis>=5.3.0
